Extract InvokeLambda to clients module
