+++ java/system-test/src/main/java/sleeper/systemtest/compaction/InvokeCompactionJobCreation.java
@@ -15,7 +15,7 @@
  */
 package sleeper.systemtest.compaction;
 
-import sleeper.systemtest.util.InvokeLambda;
+import sleeper.systemtest.util.InvokeSystemTestLambda;
 
 import java.io.IOException;
 
@@ -32,6 +32,6 @@ public class InvokeCompactionJobCreation {
             return;
         }
 
-        InvokeLambda.forInstance(args[0], COMPACTION_JOB_CREATION_LAMBDA_FUNCTION);
+        InvokeSystemTestLambda.forInstance(args[0], COMPACTION_JOB_CREATION_LAMBDA_FUNCTION);
     }
 }
